+++ Changelog.md
@@ -1,3 +1,15 @@
 ## Ignition Sensors
 
+### Ignition Sensors 2.X.X (201X-XX-XX)
+
+1. Add magnetometer
+    * [Pull request 47](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/47)
+
+1. Add IMU
+    * [Pull request 44](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/44)
+
+1. Add altimeter
+    * [Pull request 43](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/43)
+
+
 ### Ignition Sensors 1.0.0 (2019-03-01)